refactor: moved CLI functionality into its own directory; calling run directly rather than using build files in target
